\encoding{utf8}
\docType{data}
\name{forestfires}
\alias{forestfires}
\title{Description of forestfires which have occured in the northeast region of Portugal}
\format{A data matrix with 517 rows and 12 columns.}
\description{
Data are composed by 517 forest fires  that have occured in the northeast region of Portugal. These forest fires are described by the following meteorological variables: seven continuous variables (four about the FWI system: FFMC, DMC, DC, ISI and two about the meteorology: temperature and relative humidity), two integer variables relating to spatial coordinates and three binary ones indicating the presence of rain, the season (summer or not summer) and the day (weekend or not weekend).}
\examples{
data("forestfires")
}
\references{
P. Cortez and A. Morais. A data mining approach to predict forest fires using meteorological data, 2007.
}
\keyword{datasets}

