
\name{MixClusParam_continuous-class}
\alias{MixClusParam_continuous-class}
\title{MixClusParam_continuous objects}
\description{
  This class contains the margin parameters of a Gaussian distribution
  \describe{ 
    \item{mu}{mean.}
    \item{sd}{standard deviation.} 
  }
}
\examples{
new("MixClusParam_continuous")

getSlots("MixClusParam_continuous")

}