
\name{MixClusParam_integer-class}
\alias{MixClusParam_integer-class}
\title{MixClusParam_integer objects}
\description{
  This class contains the margin parameter of a Poisson distribution
  \describe{ 
    \item{beta}{mean.}
  }
}
\examples{
  new("MixClusParam_integer")
  
  getSlots("MixClusParam_integer")
  
}