
\name{MixClusParam_ordinal-class}
\alias{MixClusParam_ordinal-class}
\title{MixClusParam_ordinal objects}
\description{
  This class contains the margin parameters of a multinomial distribution
  \describe{ 
    \item{beta}{Vector of probabilities.}
  }
}
\examples{
  new("MixClusParam_ordinal")
  
  getSlots("MixClusParam_ordinal")
  
}