
\name{MixClusPrior_ordinal-class}
\alias{MixClusPrior_ordinal-class}
\title{MixClusPrior_ordinal objects}
\description{
  This class contains the hyper-parameters related to a multinomial distribution
  \describe{ 
    \item{a}{Vector of the parameters of the Dirichlet distribution.}
  }
}
\examples{
  new("MixClusPrior_ordinal")
  
  getSlots("MixClusPrior_ordinal")
  
}