\encoding{utf8}
\docType{data}
\name{simu}
\alias{simu}
\title{A simulated data set}
\format{A data matrix with 200 rows and 4 columns.}
\description{
This data set was sampled from a bi-component mixture model of Gaussian copulas. It is composed with one continuous variable, one integer variable and one binary variable.
}
\examples{
data("simu")
}

\keyword{datasets}

