\encoding{utf8}
\docType{data}
\name{wines}
\alias{wines}
\title{Description of Portuguese "Vinho Verde" wines}
\format{A data matrix with 6498 rows and 13 columns.}
\description{
Data are composed by 6497 variants of the Portuguese ``Vinho Verde" wine (1599 red wines and 4898 white wines) described by eleven physicochemical continuous variables (fixed acidity, volatile acidity, citric acidity, residual sugar, chlorides, free sulphur dioxide, total density dioxide, density, pH, sulphates, alcohol) and one integer variable (note of quality). Note that the wine 4381 because it takes outliers.
The cluster analysis should conceal the  color of the wines (column 13). 
}
\examples{
data("wines")
}
\references{
P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis. Modeling wine preferences by data mining from physicochemical properties. Decision Support Systems, 47 (4): 547-553, 2009
}
\keyword{datasets}

