\encoding{utf8}
\docType{data}
\name{simu}
\alias{simu}
\title{A simulated data set}
\format{A data matrix with 100 rows and 3 columns.}
\description{
This data set was sampled from a bi-component mixture model of Gaussian copulas with equal correlation matrices between classes. It is composed with one continuous variable, one integer variable and one binary variable.
}
\examples{
data("simu")
}

\keyword{datasets}

